% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_functions.R
\name{get_abs_paths}
\alias{get_abs_paths}
\title{Get absolute paths to important files and folders appropriate for this computer}
\usage{
get_abs_paths(
  home_path = as.character(fs::path_home()),
  cloud_storage_path = "OneDrive - University of Leeds",
  project_path = file.path(cloud_storage_path,
    "Fellowship 1960-2015 PFU database research"),
  iea_year = "2022",
  iea_folder_path = file.path(project_path, "Input Data", "External Data",
    "IEA extended energy balance data", paste("IEA", iea_year, "energy balance data")),
  iea_data_path = file.path(iea_folder_path, paste("IEA Extended Energy Balances",
    iea_year, "(TJ).csv")),
  version = "v3.0",
  input_data_path = file.path(project_path, "Input Data", "CL-PFU Data", version),
  output_data_path = file.path(project_path, "Output Data"),
  schema_path = file.path(input_data_path, "SchemaAndFKTables.xlsx"),
  fao_data_path = file.path(input_data_path, "fao_qcl_data.rds"),
  ilo_employment_data_path = file.path(input_data_path, "ilo_employment_data.rds"),
  ilo_working_hours_data_path = file.path(input_data_path, "ilo_working_hours_data.rds"),
  hmw_analysis_data_path = file.path(input_data_path, "hmw_analysis_data.xlsx"),
  amw_analysis_data_path = file.path(input_data_path, "amw_analysis_data.xlsx"),
  mw_concordance_path = file.path(input_data_path, "FAO_ISO_MW_Mapping.xlsx"),
  country_concordance_path = file.path(input_data_path, "Country_Concordance_Full.xlsx"),
  aggregation_mapping_path = file.path(input_data_path, "aggregation_mapping.xlsx"),
  phi_constants_path = file.path(input_data_path, "phi_constants.xlsx"),
  exemplar_table_path = file.path(input_data_path, "Exemplar_Table.xlsx"),
  fu_analysis_folder = file.path(input_data_path, "FU allocation data"),
  machine_data_folder = file.path(input_data_path, "Machines - Data"),
  ceda_data_folder = file.path(input_data_path, "CEDA Data"),
  exiobase_energy_flows_path = file.path(input_data_path,
    "exiobase_energy_flows_concordance.xlsx"),
  reports_source_folders = "reports",
  reports_dest_folder = file.path(output_data_path, version, "Reports"),
  pipeline_releases_folder = file.path(output_data_path, "old_pins", "PipelineReleases"),
  versions_and_products_path = file.path(pipeline_releases_folder,
    "versions and products.xlsx"),
  fsep = .Platform$file.sep
)
}
\arguments{
\item{home_path}{The absolute path to the user's home directory.}

\item{cloud_storage_path}{The path to the user's cloud storage directory, relative to \code{home_path}.}

\item{project_path}{The path to the project directory, relative to \code{home_path}.}

\item{iea_year}{The year of IEA data release, as a string.}

\item{iea_folder_path}{The path to the IEA data directory, relative to \code{home_path}.}

\item{iea_data_path}{The path to the IEA data file, relative to \code{home_path}.}

\item{version}{The version of the input data.}

\item{input_data_path}{The path to the input data directory, relative to \code{home_path}.}

\item{output_data_path}{The path to the output data directory, relative to \code{home_path}.}

\item{schema_path}{The path to the database schema file relative to \code{project_path}.
Default is}

\item{fao_data_path}{The path to the FAO live animals data file, relative to \code{home_path}.}

\item{ilo_employment_data_path}{The path to the ILO employment data file, relative to \code{home_path}.}

\item{ilo_working_hours_data_path}{The path to the ILO working hours data file, relative to \code{home_path}.}

\item{hmw_analysis_data_path}{The path to the human muscle work analysis file, relative to \code{home_path}.}

\item{amw_analysis_data_path}{The path to the animal muscle work file, relative to \code{home_path}.}

\item{mw_concordance_path}{The path to the muscle work country concordance file, relative to \code{home_path}.}

\item{country_concordance_path}{The path to the country concordance file, relative to \code{home_path}.}

\item{aggregation_mapping_path}{The path to the aggregation mapping file, relative to \code{home_path}.}

\item{phi_constants_path}{The path to the exergy-to-energy ratio file containing constant values, relative to \code{home_path}.}

\item{exemplar_table_path}{The path to the exemplar table, relative to \code{home_path}.}

\item{fu_analysis_folder}{The path to the folder containing final-to-useful exergy information, relative to \code{home_path}.}

\item{machine_data_folder}{The path to the folder containing machine-specific efficiencies, relative to \code{home_path}.}

\item{ceda_data_folder}{The path to the folder containing CEDA (Centre for Environmental Data Analysis) data, relative to \code{home_path}.}

\item{exiobase_energy_flows_path}{The path to exiobase energy flows.}

\item{reports_source_folders}{A string vector of paths to folders containing report sources.}

\item{reports_dest_folder}{The path to the folder into which reports will be written, relative to \code{home_path}.}

\item{pipeline_releases_folder}{The path to a folder containing released versions of the PSUT target data frame, relative to \code{home_path}.
Data are stored using the \code{pins} package.
Default is "PipelineReleases" relative to \code{project_path}.}

\item{versions_and_products_path}{The path to the versions and products file containing lists of all versions and products,
relative to \code{project_path}.
Default is "versions_and_products.xlsx".}

\item{fsep}{The file separator used when constructing paths.
Default is \code{.Platform$file.sep}.}
}
\value{
A named list containing paths to important directories and files.
See details for output items.
}
\description{
Default argument values assume a directory structure that includes Dropbox
and is appropriate for the Fellowship project for Dr. Paul Brockway.
}
\details{
The default argument for \code{home_path} gets the value for \code{fs::path_home()}.

Although this function is called \code{\link[=get_abs_paths]{get_abs_paths()}},
it can return relative paths if both \code{home_path} and \code{cloud_storage_path}
are set to "" (an empty string) and
\code{project_path} is set to the root folder of the project.
In that event, the leading file separator (\code{fsep}) is removed
from the path (if it exists).

Output items include:
\describe{
\item{home_path}{The absolute path to the user's home.}
\item{cloud_storage_path}{The absolute path of the user's cloud storage folder.}
\item{project_path}{The absolute path to the project folder.}
\item{iea_folder_path}{The absolute path to a folder containing IEA data.}
\item{iea_data_path}{The absolute path to the IEA data file for the OECD countries.}
\item{input_data_path}{The absolute path to the input data directory.}
\item{output_data_path}{The absolute path to the output data directory.}
\item{schema_path}{The absolute path to a Excel schema file.}
\item{fao_data_path}{The absolute path to the FAO live animals data file.}
\item{ilo_employment_data_path}{The absolute path to the ILO employment data file.}
\item{ilo_working_hours_data_path}{The absolute path to the ILO working hours data file.}
\item{hmw_analysis_data_path}{The absolute path to the human muscle work analysis file.}
\item{amw_analysis_data_path}{The absolute path to the animal muscle work file.}
\item{mw_concordance_path}{The absolute path to the muscle work country concordance file.}
\item{country_concordance_path}{The absolute path to the country concordance file.}
\item{aggregation_mapping_path}{The absolute path to the aggregation mapping file.}
\item{phi_constants_path}{The absolute path to the exergy-to-energy ratio file containing constant values.}
\item{exemplar_table_path}{The path to the exemplar table.}
\item{fu_analysis_folder}{The absolute path to the folder containing final-to-useful exergy information.}
\item{machine_data_folder}{The absolute path to the folder containing machine-specific efficiency information.}
\item{ceda_data_folder}{The absolute path to the folder containing CEDA environment information.}
\item{exiobase_energy_flows_path}{The absolute path to the folder containing Exiobase energy flows information.}
\item{reports_source_folders}{A vector of absolute paths to folders containing source reports.}
\item{reports_dest_folder}{A directory into which completed reports will be written.}
\item{pipeline_caches_folder}{The path to a folder that stores zipped versions of the pipeline cache.}
\item{pipeline_releases_folder}{The path to a folder that stores releases of various targets.}
\item{versions_and_products_path}{The path to the versions and products file.}
}
}
\examples{
get_abs_paths()
# Set relative paths with empty strings
get_abs_paths(home_path = "", cloud_storage_path = "",
              project_path = "my_project_path")
}
