% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin_functions.R
\name{pin_names}
\alias{pin_names}
\title{Get pin strings}
\usage{
pin_names(
  version,
  version_string = paste0(version_prefix, version),
  versions_and_products_path = get_abs_paths()[["versions_and_products_path"]],
  version_table_tab = "version_table",
  product_column = "product",
  pin_name_column = "pin_name",
  version_prefix = "v"
)
}
\arguments{
\item{version}{A number representing the version for which pin names should be returned.}

\item{version_string}{A string, prefixed with "v" for the version of interest.
Default is \code{paste0("v", version)}.
The \code{version_string} is used internally.}

\item{versions_and_products_path}{The path to the "versions and products.xlsx" file
that contains a concordance table of versions
and pin names.}

\item{version_table_tab}{The name of the \code{version_table} tab in the
versions and products file.
Default is "version_table".}
}
\value{
A named list of pin version strings.
Names are the various pin names or the product names.
Values are the pin version strings.
}
\description{
The CL-PFU database uses the \code{pins} package
to store results.
This function translates a CL-PFU database version number to
a pin string for later retrieval.
}
\details{
The return value is a named list where the values are the pin version strings
and the name are either product names or pin version strings.
}
\examples{
pin_names(1.1)
pin_names(version_string = "v1.1")
}
