% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin_functions.R
\name{pin_versions}
\alias{pin_versions}
\title{Get pin version strings for a CL-PFU database release}
\usage{
pin_versions(
  database_version,
  versions_and_products_path = get_abs_paths()[["versions_and_products_path"]],
  product_table_tab = "product_table",
  product_column = "product",
  pin_name_column = "pin_name"
)
}
\arguments{
\item{database_version}{A string, prefixed with "v" for the version of interest.
Any number will be prefixed by "v" and converted to a string internally.}

\item{versions_and_products_path}{The path to the "versions and products.xlsx" file
that contains a concordance table of versions
and pin names.
Default is \code{get_abs_paths()[["versions_and_products_path"]]}.}

\item{product_table_tab}{The name of the \code{version_table} tab in the
versions and products file.
Default is "version_table".}

\item{product_column}{The name of the product column in the version table.
Default is "product".}

\item{pin_name_column}{The name of the pin name column in the version table.
Default is "pin_name".}
}
\value{
A named list of pin version strings.
Names are the various pin names or the product names.
Values are the pin version strings.
}
\description{
The CL-PFU database uses the \code{pins} package
to store results.
This function gives the pin version strings for CL-PFU database release.
}
\details{
The return value is a named list where the values are the pin version strings
and the name are either product names or pin version strings.
}
\examples{
\dontrun{
pin_versions("v1.2")
pin_versions(1.2)
}
}
