% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin_functions.R
\name{read_pin_version}
\alias{read_pin_version}
\title{Read a version of a pinned CL-PFU database product}
\usage{
read_pin_version(
  pin_name,
  database_version,
  pin_version_string = pin_versions(database_version)[[pin_name]],
  pipeline_releases_folder = get_abs_paths()[["pipeline_releases_folder"]]
)
}
\arguments{
\item{pin_name}{The string name of the pin to be read.}

\item{database_version}{A string, prefixed with "v" for the version of interest.
Any number will be prefixed by "v" and converted to a string internally.}

\item{pin_version_string}{The version string for pin \code{pin_name} associated with \code{database_version}.
Default is \code{pin_versions(database_version)[[pin_name]]}.}

\item{pipeline_releases_folder}{The path to the pipeline releases folder.
Default is \code{get_abs_paths()[["pipeline_releases_folder"]]}.}
}
\value{
The pinned object represented by the name and the version string.
}
\description{
Read a version of a pinned CL-PFU database product
}
\examples{
\dontrun{
read_pin_version(pin_name = "phi_vecs", database_version = 1.2) |>
  head()
}
}
