% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_functions.R
\name{get_abs_paths}
\alias{get_abs_paths}
\title{Get absolute paths to important files and folders appropriate for this computer}
\usage{
get_abs_paths(
  home_path = fs::path_home() \%>\% as.character(),
  dropbox_path = "Dropbox",
  project_path = file.path(dropbox_path, "Fellowship 1960-2015 PFU database"),
  iea_folder_path = file.path(project_path, "IEA extended energy balance data",
    "IEA 2021 energy balance data"),
  iea_data_path = file.path(iea_folder_path,
    "IEA Extended Energy Balances 2021 (ktoe).csv"),
  version = "v1.0",
  inputData_path = file.path(project_path, "InputData", version),
  outputData_path = file.path(project_path, "OutputData"),
  fao_data_path = file.path(inputData_path, "fao_qcl_data.rds"),
  ilo_data_path = file.path(inputData_path, "ilo_hmw_data.rds"),
  hmw_analysis_data_path = file.path(inputData_path, "hmw_analysis_data.xlsx"),
  amw_analysis_data_path = file.path(inputData_path, "amw_analysis_data.xlsx"),
  mw_concordance_path = file.path(inputData_path, "FAO_ISO_MW_Mapping.xlsx"),
  country_concordance_path = file.path(inputData_path, "Country_Concordance_Full.xlsx"),
  aggregation_mapping_path = file.path(inputData_path, "aggregation_mapping.xlsx"),
  phi_constants_path = file.path(inputData_path, "Data", "Phi - Data",
    "phi_constants.xlsx"),
  exemplar_table_path = file.path(inputData_path, "Exemplar_Table.xlsx"),
  fu_analysis_folder = file.path(inputData_path, "FU analysis data"),
  ceda_data_folder = file.path(inputData_path, "CEDA Data"),
  machine_data_folder = file.path(inputData_path, "Machines - Data"),
  reports_source_folders = "reports",
  reports_dest_folder = file.path(outputData_path, "Reports"),
  pipeline_caches_folder = file.path(outputData_path, "PipelineCaches"),
  pipeline_releases_folder = file.path(outputData_path, "PipelineReleases")
)
}
\arguments{
\item{home_path}{The absolute path to the user's home directory.}

\item{dropbox_path}{The path to the user's Dropbox directory, relative to \code{home_path}.}

\item{project_path}{The path to the project directory, relative to \code{home_path}.}

\item{iea_folder_path}{The path to the IEA data directory, relative to \code{home_path}.}

\item{iea_data_path}{The path to the IEA data file, relative to \code{home_path}.}

\item{version}{The version of the inputData.}

\item{inputData_path}{The path to the inputData directory.}

\item{outputData_path}{THe path to the outputData directory.}

\item{fao_data_path}{The path to the FAO live animals data file, relative to \code{home_path}.}

\item{ilo_data_path}{The path to the ILO employment data file, relative to \code{home_path}.}

\item{hmw_analysis_data_path}{The path to the human muscle work analysis file, relative to \code{home_path}.}

\item{amw_analysis_data_path}{The path to the animal muscle work file, relative to \code{home_path}.}

\item{mw_concordance_path}{The path to the muscle work country concordance file, relative to \code{home_path}.}

\item{country_concordance_path}{The path to the country concordance file, relative to \code{home_path}.}

\item{aggregation_mapping_path}{The path to the aggregation mapping file, relative to \code{home_path}.}

\item{phi_constants_path}{The path to the exergy-to-energy ratio file containing constant values, relative to \code{home_path}.}

\item{exemplar_table_path}{The path to the exemplar table, relative to \code{home_path}.}

\item{fu_analysis_folder}{The path to the folder containing final-to-useful exergy information, relative to \code{home_path}.}

\item{ceda_data_folder}{The path to the folder containing CEDA (Centre for Environmental Data Analysis) data, relative to \code{home_path}.}

\item{machine_data_folder}{The path to the folder containing machine-specific efficiencies, relative to \code{home_path}.}

\item{reports_source_folders}{A string vector of paths to folders containing report sources.}

\item{reports_dest_folder}{The path to the folder into which reports will be written, relative to \code{home_path}.}

\item{pipeline_caches_folder}{The path to a folder containing zipped versions of the pipeline cache.
Data are stored as .zip files
Default is "PipelineCaches" relative to \code{project_path}.}

\item{pipeline_releases_folder}{The path to a folder containing released versions of the PSUT target data frame.
Data are stored using the \code{pins} package.
Default is "PipelineReleases" relative to \code{project_path}.}
}
\value{
A named list containing paths to important directories and files, including:
\itemize{
\item{\code{home_path}}{The absolute path to the user's home.}
\item{\code{dropbox_path}}{The absolute path of the user's Dropbox folder.}
\item{\code{project_path}}{The absolute path to the project folder.}
\item{\code{iea_folder_path}}{The absolute path to a folder containing IEA data.}
\item{\code{iea_data_path}}{The absolute path to the IEA data file for the OECD countries.}
\item{\code{inputData_path}}{The absolute path to the inputData directory.}
\item{\code{outputData_path}}{The absolute path to the outputData directory.}
\item{\code{fao_data_path}}{The absolute path to the FAO live animals data file.}
\item{\code{ilo_data_path}}{The absolute path to the ILO employment data file.}
\item{\code{hmw_analysis_data_path}}{The absolute path to the human muscle work analysis file.}
\item{\code{amw_analysis_data_path}}{The absolute path to the animal muscle work file.}
\item{\code{mw_concordance_path}}{The absolute path to the muscle work country concordance file.}
\item{\code{country_concordance_path}}{The absolute path to the country concordance file.}
\item{\code{aggregation_mapping_path}}{The absolute path to the aggregation mapping file.}
\item{\code{phi_constants_path}}{The absolute path to the exergy-to-energy ratio file containing constant values.}
\item{\code{exemplar_table_path}}{The path to the exemplar table.}
\item{\code{fu_analysis_folder}}{The absolute path to the folder containing final-to-useful exergy information.}
\item{\code{machine_data_folder}}{The absolute path to the folder containing machine-specific efficiency information.}
\item{\code{ceda_data_folder}}{The absolute path to the folder containing CEDA environment information.}
\item{\code{reports_source_folders}}{A vector of absolute paths to folders containing source reports.}
\item{\code{reports_dest_folder}}{A directory into which completed reports will be written.}
\item{\code{pipeline_caches_folder}}{The path to a folder that stores zipped versions of the pipeline cache.}
\item{\code{pipeline_releases_folder}}{The path to a folder that stores releases of various targets.}
}
}
\description{
Default argument values assume a directory structure that includes Dropbox
and is appropriate for the Fellowship project for Dr. Paul Brockway.
}
\details{
The default argument for \code{home_path} gets the value for \code{fs::path_home()}.
}
\examples{
get_abs_paths()
}
