% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_functions.R
\name{get_abs_paths}
\alias{get_abs_paths}
\title{Get absolute paths to important files and folders appropriate for this computer}
\usage{
get_abs_paths(
  home_path = fs::path_home(),
  dropbox_path = "Dropbox",
  project_path = file.path(dropbox_path, "Fellowship 1960-2015 PFU database"),
  iea_folder_path = file.path(project_path, "IEA extended energy balance data",
    "IEA 2019 energy balance data"),
  iea_data_path = file.path(iea_folder_path, "IEA Extended Energy Balances 2019.csv"),
  exemplar_table_path = file.path(project_path, "Database plan", "Exemplar_Table.xlsx"),
  fu_analysis_folder = file.path(project_path, "Country-level exergy accounting data"),
  ceda_data_folder = file.path(project_path, "Data", "CEDA Data"),
  machine_data_folder = file.path(project_path, "Data", "Machines - Data"),
  reports_source_folders = "reports",
  reports_dest_folder = file.path(project_path, "Reports")
)
}
\arguments{
\item{home_path}{The absolute path to the user's home directory.}

\item{dropbox_path}{The path to the user's Dropbox directory, relative to \code{home_path}.}

\item{project_path}{The path to the project directory, relative to \code{dropbox_path}.}

\item{iea_folder_path}{The path to the IEA data directory, relative to \code{project_path}.}

\item{iea_data_path}{The path to the IEA data file, relative to \code{iea_folder_path}.}

\item{exemplar_table_path}{The path to the exemplar table, relative to \code{project_path}.}

\item{fu_analysis_folder}{The path to the folder containing final-to-useful exergy information, relative to \code{project_path}.}

\item{ceda_data_folder}{The path to the folder containing CEDA (Centre for Environmental Data Analysis) data, relative to \code{project_path}.}

\item{machine_data_folder}{The path to the folder containing machine-specific efficiencies, relative to \code{project_path}.}

\item{reports_source_folders}{A string vector of paths to folders containing report sources.}

\item{reports_dest_folder}{The path to the folder into which reports will be written, relative to \code{project_path}.}
}
\value{
A named list containing paths to important directories and files, including
\code{home_path} (the absolute path to the user's home),
\code{dropbox_path} (the absolute path of the user's Dropbox folder)
\code{project_path} (the absolute path to the project folder),
\code{iea_folder_path} (the absolute path to a folder containing IEA data),
\code{iea_data_path} (the absolute path to the IEA data file for the OECD countries),
\code{fu_analysis_path} (the absolute path to the folder containing final-to-useful exergy information),
\code{machine_data_folder} (the absolute path to the folder containing machine-specific efficiency information),
\code{ceda_data_folder} (the absolute path to the folder containing CEDA environment information),
\code{reports_source_folders} (a vector of absolute paths to folders containing source reports), and
\code{reports_dest_folder} (a directory into which completed reports will be written).
}
\description{
Default argument values assume a directory structure that includes Dropbox
and is appropriate for the Fellowship project for Dr. Paul Brockway.
}
\details{
The default argument for \code{home_path} gets the value for \code{Sys.getenv("HOME")}.
On Windows, user directories reported by Sys.getenv("HOME") are set to the user's "Documents" folder by default.
\code{R} on other OSs does not append the "Documents" folder at the end of the \code{home_path}.
For the \code{home_path} argument,
we want the path without the Documents folder appended, because that is the enclosing directory
for the Dropbox folder's default location.
Thus, the default argument for \code{home_path} trims trailing "Documents", if present.
The default value for the \code{home_path} argument won't find "Documents" at the end of macOS and Linux home paths
and will return the HOME path, as desired.
}
\examples{
get_abs_paths()
}
