% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_functions.R
\name{get_abs_paths}
\alias{get_abs_paths}
\title{Get absolute paths to important files and folders appropriate for this computer}
\usage{
get_abs_paths(
  home_path = fs::path_home() \%>\% as.character(),
  dropbox_path = "Dropbox",
  project_path = file.path(dropbox_path, "Fellowship 1960-2015 PFU database"),
  iea_folder_path = file.path(project_path, "IEA extended energy balance data",
    "IEA 2019 energy balance data"),
  iea_data_path = file.path(iea_folder_path, "IEA Extended Energy Balances 2019.csv"),
  country_concordance_path = file.path(project_path, "Mapping",
    "Country_Concordance_Full.xlsx"),
  exemplar_table_path = file.path(project_path, "Database plan", "Exemplar_Table.xlsx"),
  fu_analysis_folder = file.path(project_path, "Country-level exergy accounting data"),
  ceda_data_folder = file.path(project_path, "Data", "CEDA Data"),
  machine_data_folder = file.path(project_path, "Data", "Machines - Data"),
  reports_source_folders = "reports",
  reports_dest_folder = file.path(project_path, "Reports"),
  drake_cache_folder = file.path(getwd(), "..", "PFU-Database", ".drake")
)
}
\arguments{
\item{home_path}{The absolute path to the user's home directory.}

\item{dropbox_path}{The path to the user's Dropbox directory, relative to \code{home_path}.}

\item{project_path}{The path to the project directory, relative to \code{dropbox_path}.}

\item{iea_folder_path}{The path to the IEA data directory, relative to \code{project_path}.}

\item{iea_data_path}{The path to the IEA data file, relative to \code{iea_folder_path}.}

\item{country_concordance_path}{The path to the country concordance file, relative to \code{project_path}.}

\item{exemplar_table_path}{The path to the exemplar table, relative to \code{project_path}.}

\item{fu_analysis_folder}{The path to the folder containing final-to-useful exergy information, relative to \code{project_path}.}

\item{ceda_data_folder}{The path to the folder containing CEDA (Centre for Environmental Data Analysis) data, relative to \code{project_path}.}

\item{machine_data_folder}{The path to the folder containing machine-specific efficiencies, relative to \code{project_path}.}

\item{reports_source_folders}{A string vector of paths to folders containing report sources.}

\item{reports_dest_folder}{The path to the folder into which reports will be written, relative to \code{project_path}.}

\item{drake_cache_folder}{The path to the drake cache. Default is \code{../PFU-Database/.drake}.
See Details for more information.}
}
\value{
A named list containing paths to important directories and files, including
\itemize{
\item{\code{home_path}}{The absolute path to the user's home.}
\item{\code{dropbox_path}}{The absolute path of the user's Dropbox folder.}
\item{\code{project_path}}{The absolute path to the project folder.}
\item{\code{iea_folder_path}}{The absolute path to a folder containing IEA data.}
\item{\code{iea_data_path}}{The absolute path to the IEA data file for the OECD countries.}
\item{\code{country_concordance_path}}{The absolute path to the country concordance file.}
\item{\code{exemplar_table_path}}{The path to the exemplar table.}
\item{\code{fu_analysis_folder}}{The absolute path to the folder containing final-to-useful exergy information.}
\item{\code{machine_data_folder}}{The absolute path to the folder containing machine-specific efficiency information.}
\item{\code{ceda_data_folder}}{The absolute path to the folder containing CEDA environment information.}
\item{\code{reports_source_folders}}{A vector of absolute paths to folders containing source reports.}
\item{\code{reports_dest_folder}}{A directory into which completed reports will be written.}
\item{\code{drake_cache_folder}}{The path to the drake cache.}
}
}
\description{
Default argument values assume a directory structure that includes Dropbox
and is appropriate for the Fellowship project for Dr. Paul Brockway.
}
\details{
The default argument for \code{home_path} gets the value for \code{fs::path_home()}.

The default value for \code{drake_cache_folder} assumes that the \code{PFUSetup} folder
(this project's folder) is a sibling of the working directory
from which the drake workflow is run.
If this assumption is not true, users are free to
pass a non-default value into this function.
}
\examples{
get_abs_paths()
}
